refactor(kitchenpos): refactoring package by context
