+++ src/main/java/kitchenpos/domain/order/OrderTableRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.order;
 
 import java.util.List;
 import java.util.Optional;
