+++ src/main/java/kitchenpos/domain/order/OrderType.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.order;
 
 public enum OrderType {
     DELIVERY, TAKEOUT, EAT_IN
