+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -1,13 +1,10 @@
 package kitchenpos.application;
 
-import kitchenpos.domain.MenuRepository;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.OrderRepository;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.OrderTableRepository;
-import kitchenpos.domain.OrderType;
+import kitchenpos.domain.menu.InMemoryMenuRepository;
+import kitchenpos.domain.order.*;
+import kitchenpos.infra.FakeKitchenridersClient;
+import kitchenpos.domain.menu.MenuRepository;
+import kitchenpos.application.OrderService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
