+++ src/test/java/kitchenpos/domain/menu/InMemoryMenuRepository.java
@@ -1,7 +1,4 @@
-package kitchenpos.application;
-
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuRepository;
+package kitchenpos.domain.menu;
 
 import java.util.ArrayList;
 import java.util.HashMap;
