+++ src/test/java/kitchenpos/domain/order/InMemoryOrderRepository.java
@@ -1,9 +1,4 @@
-package kitchenpos.application;
-
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderRepository;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
+package kitchenpos.domain.order;
 
 import java.util.ArrayList;
 import java.util.HashMap;
