+++ src/test/java/kitchenpos/domain/order/InMemoryOrderTableRepository.java
@@ -1,7 +1,4 @@
-package kitchenpos.application;
-
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.OrderTableRepository;
+package kitchenpos.domain.order;
 
 import java.util.ArrayList;
 import java.util.HashMap;
