+++ src/test/java/kitchenpos/domain/product/InMemoryProductRepository.java
@@ -1,7 +1,4 @@
-package kitchenpos.application;
-
-import kitchenpos.domain.Product;
-import kitchenpos.domain.ProductRepository;
+package kitchenpos.domain.product;
 
 import java.util.ArrayList;
 import java.util.HashMap;
